# backend/src/factors/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Factors usage doc for golang api server.",
        "title": "Factors Backend Api",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "factors-dev.com:8080",
    "basePath": "/projects",
    "paths": {
...
        "/{project_id}/models": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factors"
                ],
                "summary": "To get model infos for the given project id.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"mid\": uint64, \"mt\": string, \"st\": timestamp, \"et\": timestamp}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/{project_id}/v1/event_names": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V1Api"
                ],
                "summary": "Get event names for the given project id.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"event_names\": map[string][]string}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}